refactor: 잔돈 충전 부분 분리
