+++ src/pages/MainPage/index.tsx
@@ -1,3 +1,4 @@
+import ChangeInput from 'components/ChangeInput';
 import Menu from 'components/Menu';
 import { useRecoilState } from 'recoil';
 import { changeState, dispenserState, guideState } from 'recoil/states';
@@ -11,23 +12,6 @@ function MainPage() {
   const [change, setChange] = useRecoilState(changeState);
   const [dispenser, setDispenser] = useRecoilState(dispenserState);
 
-  const chargeChange = (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!(e.target instanceof HTMLFormElement)) return;
-
-    const money = (e.target.elements[0] as HTMLInputElement).valueAsNumber;
-
-    if (!money || money % 500 !== 0) {
-      alert('500원 단위로 입력해주세요.');
-      return;
-    }
-
-    setChange(prevState => prevState + money);
-
-    setGuiding('투입되었습니다.');
-  };
-
   const returnChange = () => {
     if (!window.confirm('반환하시겠습니까?')) return;
 
@@ -40,10 +24,7 @@ function MainPage() {
     <S.VendingMachine>
       <Menu />
       <S.Guiding>{guiding}</S.Guiding>
-      <S.MoneyInputForm onSubmit={chargeChange}>
-        <input type="number" />
-        <button>투입</button>
-      </S.MoneyInputForm>
+      <ChangeInput />
       <S.ChangeContainer>
         <div>{change.toLocaleString('ko-KR')}</div>
         <button type="button" onClick={returnChange}>
